Two bug fixes for `unescape()`
### Incorrectly handled special code points

Previously, `unescape()` outputted the replacement character (U+FFFD
“�”) for certain numeric entities:

  * [Noncharacters]
  * [Control] characters
  * `0x0D` (carriage return)

A close reading of the [spec] and some browser testing shows that
behavior to be incorrect. Those characters are now outputted as
themselves.

### Incorrectly handled long numeric entities

`unescape()` incorrectly outputted long numeric entities as the literal
text of the entity.

A close reading of the [spec] and some browser testing shows that
behavior to be incorrect. Those long entities are now outputted as the
replacement character (U+FFFD “�”).

[Noncharacters]: https://infra.spec.whatwg.org/#noncharacter
[Control]: https://infra.spec.whatwg.org/#control
[spec]: https://html.spec.whatwg.org/multipage/parsing.html#numeric-character-reference-end-state