Improve documentation for escape functions
This adds examples for each escape function, tries to clarify when they
should be used, and simplifies the table showing which functions escape
which characters.

This also removes the warning about comments, since [according to the
spec], escaping < and > should prevent the early termination of an HTML
comment. (I had though that `--` was an issue, but apparently not.)

[according to the spec]: https://html.spec.whatwg.org/multipage/syntax.html#comments