escape: Separate finding and mapping special characters
This separates finding the position of epecial characters in an input
slice from mapping them to an entity and adding them to the output
string.

Strictly speaking we could have the function that finds the special
characters return the correct entity, but this seems to be slower. The
performance issues show up even in in benchmarks that seem like they
shouldn’t be affected, e.g. with clean inputs, so this may be some sort
of optimizer wonkiness.