Refactor and update `escape` benchmark inputs
Previously, the “small” inputs were a fragment of HTML, but I think the
“Björk and Борис” test string is both better exercise (because of all
the multibyte characters) and more realistic (fewer characters need to
be escaped).

This adds a simple “medium” test case with only three special
characters, but of medium length.

This also adds the clean version of the “Björk and Борис” string to the
regular `escape` tests to confirm that they don’t change it.