unescape: clean up `correct_numeric_entity()`
This updates `correct_numeric_entity()` to return `Cow<[u8]>` instead of
`Option<Vec<u8>>`.

It brings a few of the test functions (`is_outside_range()` and
`is_surrogate()`) into the main `match` statement since they are quite
simple, and it handles the case where `char::from_u32()` returned `None`
(shouldn’t happen).

It also adds a few tests to make sure that the right variant of `Cow` is
being returned — it would be a shame to have it allocating when I went
to the trouble of using `Cow`.