Add license file for entities.json
I’m unclear on whether the entities.json file can actually be
copyrighted, but if it can be then it appears to be owned by WHATWG, and
they license it under the BSD 3-Clause License.

This adds an entities.json-LICENSE file and updates the
update-entities.sh script to update the license file. It also clarifies
the situation in README.md.

Fixes: #58 Reference or include license for entities.json