Faster entity unescaping in phf (slow) mode
Previously we found a candidate entity and checked it for entity matches
from longest to shortest. This changes the algorithm to first check for
a long match if the candidate ends with a semicolon, then check the
shortest possible match to the longest possible.

This works because no bare entity (without a semicolon) is a prefix for
any other bare entity, so if a short candidate matches, then there can
be no matching longer candidate.